fix: Fix overflow on 32-bit platforms when writing big files.
Not all definitions are changed, because for example the `Read` and
`Write` calls are really dealing with `size_t` since they are limited
by what you can keep in memory. While it wouldn't be wrong to use
`int64_t` in these cases, they are a deeply entrenched API which is
also used in the client, so I'd rather not change those.

Changelog: Title
Ticket: MEN-7613

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>