fix: Fix handling of the '-f' short option
`getopt_long()` is called in two places in the code, so all
options need to be present in both of the calls. Let's avoid this
in the future by defining a macro and using it in both places.

Ticket: none
Changelog: Fixed handling of the '-f' short option for '--fsync-interval'
Signed-off-by: Vratislav Podzimek <vratislav.podzimek+auto-signed@northern.tech>